Install a custom version of Ruby if the system doesn't have it
By changing the Gemfile depends recently to include a sitemap and a specific
version of Jekyll that fixes some security holes, we now need a newer version
of Ruby to build Jekyll.

To deal with this, I've made a requirement for a very recent version of Ruby
(if you're going to upgrade, go as new as you can)
and use RVM to install it if the system doesn't provide it.

This *should* all be automatic. By default it will attempt to compile Ruby
rather than use system packages. #TODO: use system packages when available.

This also includes a modified envrc sample that loads the RVM ruby version
and sets some other environment variables.

In addition, the _config.yml for Jekyll now ONLY EXPLICITLY INCLUDES FILES
TO BUILD. This is in part to avoid trying to build rvm/ files, and also
an intent to only spend time building the files we need.

TODO:
 - try to use system packages to install Ruby version